# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AUTH API",
        "description": "AUTH API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/permission/": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get a list of permissions",
                "description": "Get a list of permissions",
                "operationId": "c665d12b29fb188041ae697110acd3fb",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Id of the permission",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Name of the permission",
                                                "type": "string"
                                            },
                                            "guard_name": {
                                                "description": "name of the guard",
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "description": "created date of the permission",
                                                "type": "string"
                                            },
                                            "updated_at": {
                                                "description": "updated date of the permission",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Create a permission",
                "description": "Create a permission",
                "operationId": "20223fd34738a7a9cd8a27e37a0b8daf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "worker"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        },
                                        "permission": {
                                            "properties": {
                                                "id": {
                                                    "description": "Id of the permission",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Name of the permission",
                                                    "type": "string"
                                                },
                                                "guard_name": {
                                                    "description": "name of the guard",
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "description": "created date of the permission",
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "description": "updated date of the permission",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}